Rename annotation servcie to Notes service
